remove AbstractMachineAllocationController level
