+++ spinn_front_end_common/abstract_models/__init__.py
@@ -15,8 +15,6 @@
 from .abstract_generates_data_specification import (
     AbstractGeneratesDataSpecification)
 from .abstract_has_associated_binary import AbstractHasAssociatedBinary
-from .abstract_machine_allocation_controller import (
-    AbstractMachineAllocationController)
 from .abstract_rewrites_data_specification import (
     AbstractRewritesDataSpecification)
 from .abstract_send_me_multicast_commands_vertex import (
@@ -34,7 +32,6 @@ from .has_custom_atom_key_map import HasCustomAtomKeyMap
 
 __all__ = ["AbstractGeneratesDataSpecification",
            "AbstractHasAssociatedBinary",
-           "AbstractMachineAllocationController",
            "AbstractRewritesDataSpecification",
            "AbstractSendMeMulticastCommandsVertex",
            "AbstractSupportsDatabaseInjection",
