+++ spinn_front_end_common/abstract_models/impl/machine_allocation_controller.py
@@ -16,19 +16,15 @@ import logging
 import sys
 from threading import Thread
 from spinn_utilities.log import FormatAdapter
-from spinn_utilities.overrides import overrides
 from spinn_utilities.abstract_base import AbstractBase, abstractmethod
 from spinnman.constants import SCP_SCAMP_PORT
 from spinnman.connections.udp_packet_connections import SCAMPConnection
 from spinnman.transceiver import create_transceiver_from_hostname
-from spinn_front_end_common.abstract_models import (
-    AbstractMachineAllocationController)
 from spinnman.connections.udp_packet_connections import EIEIOConnection
 logger = FormatAdapter(logging.getLogger(__name__))
 
 
-class MachineAllocationController(
-        AbstractMachineAllocationController, metaclass=AbstractBase):
+class MachineAllocationController(object, metaclass=AbstractBase):
     """
     How to manage the allocation of a machine so that it gets cleaned up
     neatly when the script dies.
@@ -52,8 +48,20 @@ class MachineAllocationController(
         self.__connection_data = connection_data
         thread.start()
 
-    @overrides(AbstractMachineAllocationController.close)
+    @abstractmethod
+    def extend_allocation(self, new_total_run_time):
+        """
+        Extend the allocation of the machine from the original run time.
+
+        :param float new_total_run_time:
+            The total run time that is now required starting from when the
+            machine was first allocated
+        """
+
     def close(self):
+        """
+        Indicate that the use of the machine is complete.
+        """
         self._exited = True
 
     @abstractmethod
@@ -64,6 +72,18 @@ class MachineAllocationController(
 
         :return: Whether the machine is still (believed to be) allocated.
         :rtype: bool
+        """
+
+    @abstractmethod
+    def where_is_machine(self, chip_x, chip_y):
+        """
+        Locates and returns cabinet, frame, board for a given chip in a
+        machine allocated to this job.
+
+        :param int chip_x: chip x location
+        :param int chip_y: chip y location
+        :return: (cabinet, frame, board)
+        :rtype: tuple(int,int,int)
         """
 
     def _teardown(self):
@@ -83,8 +103,14 @@ class MachineAllocationController(
                 " the script; this script will now exit")
             sys.exit(1)
 
-    @overrides(AbstractMachineAllocationController.create_transceiver)
     def create_transceiver(self):
+        """
+        Create a transceiver for talking to the allocated machine, and
+        make sure everything is ready for use (i.e. boot and discover
+        connections if needed).
+
+        :rtype: ~spinnman.transceiver.Transceiver
+        """
         if not self.__hostname:
             return None
         txrx = create_transceiver_from_hostname(
@@ -94,8 +120,24 @@ class MachineAllocationController(
         txrx.discover_scamp_connections()
         return txrx
 
-    @overrides(AbstractMachineAllocationController.open_sdp_connection)
+    def can_create_transceiver(self):
+        return self.__hostname is not None
+
     def open_sdp_connection(self, chip_x, chip_y, udp_port=SCP_SCAMP_PORT):
+        """
+        Open a connection to a specific Ethernet-enabled SpiNNaker chip.
+        Caller will have to arrange for SpiNNaker to pay attention to the
+        connection.
+
+        The coordinates will be job-relative.
+
+        :param int chip_x: Ethernet-enabled chip X coordinate
+        :param int chip_y: Ethernet-enabled chip Y coordinate
+        :param int udp_port:
+            the UDP port on the chip to connect to; connecting to a non-SCP
+            port will result in a connection that can't easily be configured.
+        :rtype: ~spinnman.connections.udp_packet_connections.SDPConnection
+        """
         if not self.__connection_data:
             return None
         host = self.__connection_data[chip_x, chip_y]
@@ -105,8 +147,13 @@ class MachineAllocationController(
             chip_x=chip_x, chip_y=chip_y,
             remote_host=host, remote_port=udp_port)
 
-    @overrides(AbstractMachineAllocationController.open_eieio_connection)
     def open_eieio_connection(self, chip_x, chip_y):
+        """
+        Open an unbound EIEIO connection. This may be used to communicate with
+        any board of the job.
+
+        :rtype: ~spinnman.connections.udp_packet_connections.EIEIOConnection
+        """
         if not self.__connection_data:
             return None
         host = self.__connection_data[chip_x, chip_y]
@@ -114,6 +161,27 @@ class MachineAllocationController(
             return None
         return EIEIOConnection(remote_host=host, remote_port=SCP_SCAMP_PORT)
 
-    @overrides(AbstractMachineAllocationController.open_eieio_listener)
     def open_eieio_listener(self):
+        """
+        Open an unbound EIEIO connection. This may be used to communicate with
+        any board of the job.
+
+        :rtype: ~spinnman.connections.udp_packet_connections.EIEIOConnection
+        """
         return EIEIOConnection()
+
+    @property
+    def proxying(self):
+        """
+        Whether this is a proxying connection. False unless overridden.
+
+        :rtype: bool
+        """
+        return False
+
+    @abstractmethod
+    def make_report(self, filename):
+        """
+        Asks the controller to make a report of details of allocations.
+        By default, this does nothing.
+        """
