+++ spinn_front_end_common/data/fec_data_view.py
@@ -232,7 +232,7 @@ class FecDataView(PacmanDataView, SpiNNManDataView):
         """
         Returns the allocation controller if known.
 
-        :rtype: AbstractMachineAllocationController
+        :rtype: MachineAllocationController
         :raises ~spinn_utilities.exceptions.SpiNNUtilsException:
             If the buffer manager unavailable
         """
