+++ spinn_front_end_common/interface/interface_functions/hbp_allocator.py
@@ -18,8 +18,6 @@ from spinn_utilities.config_holder import get_config_str
 from spinn_utilities.overrides import overrides
 from spinn_front_end_common.abstract_models.impl import (
     MachineAllocationController)
-from spinn_front_end_common.abstract_models import (
-    AbstractMachineAllocationController)
 from spinn_front_end_common.data import FecDataView
 from pacman.exceptions import PacmanConfigurationException
 
@@ -54,7 +52,7 @@ class _HBPJobController(MachineAllocationController):
         logging.getLogger("requests").setLevel(logging.WARNING)
         super().__init__("HBPJobController")
 
-    @overrides(AbstractMachineAllocationController.extend_allocation)
+    @overrides(MachineAllocationController.extend_allocation)
     def extend_allocation(self, new_total_run_time):
         r = requests.get(self._extend_lease_url, params={
             "runTime": new_total_run_time}, timeout=10)
@@ -83,7 +81,7 @@ class _HBPJobController(MachineAllocationController):
         r.raise_for_status()
         return r.json()
 
-    @overrides(AbstractMachineAllocationController.close)
+    @overrides(MachineAllocationController.close)
     def close(self):
         super().close()
         self._release(self._machine_name)
@@ -100,7 +98,7 @@ class _HBPJobController(MachineAllocationController):
         self._set_power(self._machine_name, power)
         self._power_on = power
 
-    @overrides(AbstractMachineAllocationController.where_is_machine)
+    @overrides(MachineAllocationController.where_is_machine)
     def where_is_machine(self, chip_x, chip_y):
         return self._where_is(self._machine_name, chip_x, chip_y)
